requirements.txt
antlr4-python3-runtime==4.8
hydra-core==1.1.1
numpy==1.21.2
omegaconf==2.1.1
Pillow==8.3.2
PyYAML==5.4.1
torch==1.10.0
torchvision
typing-extensions==3.10.0.2
matplotlib
tensorboard==2.6.0
torch-scatter
torch-sparse
torch-geometric 
hydra-optuna-sweeper
statsmodels
plotly
seaborn
